+++ src/array/array_async_writable.rs
@@ -3,6 +3,7 @@ use std::{borrow::Cow, sync::Arc};
 use futures::{StreamExt, TryStreamExt};
 
 use crate::{
+    array::validate_element_size,
     array_subset::ArraySubset,
     metadata::MetadataEraseVersion,
     storage::{
@@ -264,11 +265,10 @@ impl<TStorage: ?Sized + AsyncWritableStorageTraits + 'static> Array<TStorage> {
         chunk_elements: &[T],
         options: &CodecOptions,
     ) -> Result<(), ArrayError> {
-        array_async_store_elements!(
-            self,
-            chunk_elements,
-            async_store_chunk_opt(chunk_indices, &chunk_elements, options)
-        )
+        validate_element_size::<T>(self.data_type())?;
+        let chunk_elements = crate::array::convert_to_bytes_vec(chunk_elements);
+        self.async_store_chunk_opt(chunk_indices, &chunk_elements, options)
+            .await
     }
 
     #[cfg(feature = "ndarray")]
@@ -284,14 +284,13 @@ impl<TStorage: ?Sized + AsyncWritableStorageTraits + 'static> Array<TStorage> {
         chunk_array: TArray,
         options: &CodecOptions,
     ) -> Result<(), ArrayError> {
+        validate_element_size::<T>(self.data_type())?;
         let chunk_array: ndarray::Array<T, D> = chunk_array.into();
         let chunk_shape = self.chunk_shape_usize(chunk_indices)?;
         if chunk_array.shape() == chunk_shape {
-            array_async_store_ndarray!(
-                self,
-                chunk_array,
-                async_store_chunk_elements_opt(chunk_indices, &chunk_array, options)
-            )
+            let chunk_array = super::ndarray_into_vec(chunk_array);
+            self.async_store_chunk_elements_opt(chunk_indices, &chunk_array, options)
+                .await
         } else {
             Err(ArrayError::InvalidDataShape(
                 chunk_array.shape().to_vec(),
@@ -386,11 +385,10 @@ impl<TStorage: ?Sized + AsyncWritableStorageTraits + 'static> Array<TStorage> {
         chunks_elements: &[T],
         options: &CodecOptions,
     ) -> Result<(), ArrayError> {
-        array_async_store_elements!(
-            self,
-            chunks_elements,
-            async_store_chunks_opt(chunks, &chunks_elements, options)
-        )
+        validate_element_size::<T>(self.data_type())?;
+        let chunks_elements = crate::array::convert_to_bytes_vec(chunks_elements);
+        self.async_store_chunks_opt(chunks, &chunks_elements, options)
+            .await
     }
 
     #[cfg(feature = "ndarray")]
@@ -406,15 +404,14 @@ impl<TStorage: ?Sized + AsyncWritableStorageTraits + 'static> Array<TStorage> {
         chunks_array: TArray,
         options: &CodecOptions,
     ) -> Result<(), ArrayError> {
+        validate_element_size::<T>(self.data_type())?;
         let chunks_array: ndarray::Array<T, D> = chunks_array.into();
         let chunks_subset = self.chunks_subset(chunks)?;
         let chunks_shape = chunks_subset.shape_usize();
         if chunks_array.shape() == chunks_shape {
-            array_async_store_ndarray!(
-                self,
-                chunks_array,
-                async_store_chunks_elements_opt(chunks, &chunks_array, options)
-            )
+            let chunks_array = super::ndarray_into_vec(chunks_array);
+            self.async_store_chunks_elements_opt(chunks, &chunks_array, options)
+                .await
         } else {
             Err(ArrayError::InvalidDataShape(
                 chunks_array.shape().to_vec(),