+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
  - Implement `From<&String>` for `DimensionName`
+ - Cleanup macro usage in array
 
 ## [0.15.0] - 2024-07-07